Moved gardens chart to dedicated controller
